Abundance of Fungal Communities in Cheese

Fungal community abundance of 74 ASVs sampled from the surface of three different French cheeses.

Usage

data("cheese_abundance", package = "scimo")

data("cheese_taxonomy", package = "scimo")

Format

For cheese_abundance, a tibble with columns:

sample

Sample ID.

cheese

Appellation of the cheese. One of Saint-Nectaire, Livarot or Epoisses.

rind_type

One of Natural or Washed.

other columns

Count of the ASV.

For cheese_taxonomy, a tibble with columns:

asv

Amplicon Sequence Variant (ASV) ID.

lineage

Character corresponding to a standard concatenation of taxonomic clades.

other columns

Clade to which the ASV belongs.

Source

This dataset came from doi:10.24072/pcjournal.321 .

Examples

data("cheese_abundance", package = "scimo")
cheese_abundance
#> # A tibble: 9 × 77
#>   sample    cheese    rind_type asv_01 asv_02 asv_03 asv_04 asv_05 asv_06 asv_07
#>   <chr>     <chr>     <chr>      <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>  <dbl>
#> 1 sample1-1 Saint-Ne… Natural        1      0     38     40      1      2     31
#> 2 sample1-2 Saint-Ne… Natural        3      4     38     61      4      4     48
#> 3 sample1-3 Saint-Ne… Natural       28     16     33     23     31     29     21
#> 4 sample2-1 Livarot   Washed         0      2      1      0      5      1      0
#> 5 sample2-2 Livarot   Washed         0      0      4      0      1      1      2
#> 6 sample2-3 Livarot   Washed         0      1      2      0      2      1      0
#> 7 sample3-1 Epoisses  Washed         4      2      3      0      2      5      0
#> 8 sample3-2 Epoisses  Washed         0      0      0      0      0      0      0
#> 9 sample3-3 Epoisses  Washed         0      0      1      0      0      0      2
#> # ℹ 67 more variables: asv_08 <dbl>, asv_09 <dbl>, asv_10 <dbl>, asv_11 <dbl>,
#> #   asv_12 <dbl>, asv_13 <dbl>, asv_14 <dbl>, asv_15 <dbl>, asv_16 <dbl>,
#> #   asv_17 <dbl>, asv_18 <dbl>, asv_19 <dbl>, asv_20 <dbl>, asv_21 <dbl>,
#> #   asv_22 <dbl>, asv_23 <dbl>, asv_24 <dbl>, asv_25 <dbl>, asv_26 <dbl>,
#> #   asv_27 <dbl>, asv_28 <dbl>, asv_29 <dbl>, asv_30 <dbl>, asv_31 <dbl>,
#> #   asv_32 <dbl>, asv_33 <dbl>, asv_34 <dbl>, asv_35 <dbl>, asv_36 <dbl>,
#> #   asv_37 <dbl>, asv_38 <dbl>, asv_39 <dbl>, asv_40 <dbl>, asv_41 <dbl>, …
data("cheese_taxonomy", package = "scimo")
cheese_taxonomy
#> # A tibble: 74 × 9
#>    asv    lineage                kingdom phylum class order family genus species
#>    <chr>  <chr>                  <chr>   <chr>  <chr> <chr> <chr>  <chr> <chr>  
#>  1 asv_01 k__Fungi|p__Ascomycot… Fungi   Ascom… Doth… Doth… Dothi… Aure… Aureob…
#>  2 asv_02 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Aspe… Asperg…
#>  3 asv_03 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Peni… Penici…
#>  4 asv_04 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Peni… Penici…
#>  5 asv_05 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Peni… Penici…
#>  6 asv_06 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Peni… Penici…
#>  7 asv_07 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Peni… Penici…
#>  8 asv_08 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Peni… Penici…
#>  9 asv_09 k__Fungi|p__Ascomycot… Fungi   Ascom… Euro… Euro… Asper… Peni… Penici…
#> 10 asv_10 k__Fungi|p__Ascomycot… Fungi   Ascom… Sacc… Sacc… Debar… Deba… Debary…
#> # ℹ 64 more rows