Feature aggregation step based on a hierarchical clustering

Aggregate variables according to hierarchical clustering.

Usage

step_aggregate_hclust(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  n_clusters,
  fun_agg,
  dist_metric = "euclidean",
  linkage_method = "complete",
  res = NULL,
  prefix = "cl_",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("aggregate_hclust")
)

# S3 method for step_aggregate_hclust
tidy(x, ...)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

n_clusters

Number of cluster to create.

fun_agg

Aggregation function like sum or mean.

dist_metric

Default to euclidean. See stats::dist() for more details.

linkage_method

Default to complete. See stats::hclust() for more details.

res

This parameter is only produced after the recipe has been trained.

prefix

A character string for the prefix of the resulting new variables.

keep_original_cols

A logical to keep the original variables in the output. Defaults to FALSE.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

x

A step_aggregate_hclust object.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Author

Antoine Bichat

Examples

rec <-
  iris %>%
  recipe(formula = Species ~ .) %>%
  step_aggregate_hclust(all_numeric_predictors(),
                        n_clusters = 2, fun_agg = sum) %>%
  prep()
rec
#> 
#> ── Recipe ──────────────────────────────────────────────────────────────────────
#> 
#> ── Inputs 
#> Number of variables by role
#> outcome:   1
#> predictor: 4
#> 
#> ── Training information 
#> Training data contained 150 data points and no incomplete rows.
#> 
#> ── Operations 
#> • `hclust` aggregation of: Sepal.Length and Sepal.Width, ... | Trained
tidy(rec, 1)
#> # A tibble: 4 × 3
#>   terms        aggregate id                    
#>   <chr>        <chr>     <chr>                 
#> 1 Sepal.Length cl_1      aggregate_hclust_SwlKL
#> 2 Sepal.Width  cl_2      aggregate_hclust_SwlKL
#> 3 Petal.Length cl_2      aggregate_hclust_SwlKL
#> 4 Petal.Width  cl_2      aggregate_hclust_SwlKL
juice(rec)
#> # A tibble: 150 × 3
#>    Species  cl_1  cl_2
#>    <fct>   <dbl> <dbl>
#>  1 setosa    5.1   5.1
#>  2 setosa    4.9   4.6
#>  3 setosa    4.7   4.7
#>  4 setosa    4.6   4.8
#>  5 setosa    5     5.2
#>  6 setosa    5.4   6  
#>  7 setosa    4.6   5.1
#>  8 setosa    5     5.1
#>  9 setosa    4.4   4.5
#> 10 setosa    4.9   4.7
#> # ℹ 140 more rows