Feature aggregation step based on a defined list

Aggregate variables according to prior knowledge.

Usage

step_aggregate_list(
  recipe,
  ...,
  role = "predictor",
  trained = FALSE,
  list_agg = NULL,
  fun_agg = NULL,
  others = "discard",
  name_others = "others",
  res = NULL,
  prefix = "agg_",
  keep_original_cols = FALSE,
  skip = FALSE,
  id = rand_id("aggregate_list")
)

# S3 method for step_aggregate_list
tidy(x, ...)

Arguments

recipe

A recipe object. The step will be added to the sequence of operations for this recipe.

...

One or more selector functions to choose variables for this step. See selections() for more details.

role

For model terms created by this step, what analysis role should they be assigned? By default, the new columns created by this step from the original variables will be used as predictors in a model.

trained

A logical to indicate if the quantities for preprocessing have been estimated.

list_agg

Named list of aggregated variables.

fun_agg

Aggregation function like sum or mean.

others

Behavior with selected behavior that are not present in list_agg. If discard (the default), they are not kept. If asis, they are kept without modification. If aggregate, they are aggregated in a new variable.

name_others

If others is set to aggregate, name of the aggregated variable. Not used otherwise.

res

This parameter is only produced after the recipe has been trained.

prefix

A character string for the prefix of the resulting new variables that are not named in list_agg.

keep_original_cols

A logical to keep the original variables in the output. Defaults to FALSE.

skip

A logical. Should the step be skipped when the recipe is baked by bake()? While all operations are baked when prep() is run, some operations may not be able to be conducted on new data (e.g. processing the outcome variable(s)). Care should be taken when using skip = TRUE as it may affect the computations for subsequent operations.

id

A character string that is unique to this step to identify it.

x

A step_aggregate_list object.

Value

An updated version of recipe with the new step added to the sequence of any existing operations.

Author

Antoine Bichat

Examples

list_iris <- list(sepal.size = c("Sepal.Length", "Sepal.Width"),
                  petal.size = c("Petal.Length", "Petal.Width"))
rec <-
  iris %>%
  recipe(formula = Species ~ .) %>%
  step_aggregate_list(all_numeric_predictors(),
                      list_agg = list_iris, fun_agg = prod) %>%
  prep()
rec
#> 
#> ── Recipe ──────────────────────────────────────────────────────────────────────
#> 
#> ── Inputs 
#> Number of variables by role
#> outcome:   1
#> predictor: 4
#> 
#> ── Training information 
#> Training data contained 150 data points and no incomplete rows.
#> 
#> ── Operations 
#> • Aggregation of: Sepal.Length, Sepal.Width, Petal.Length, ... | Trained
tidy(rec, 1)
#> # A tibble: 4 × 3
#>   terms        aggregate  id                  
#>   <chr>        <chr>      <chr>               
#> 1 Sepal.Length sepal.size aggregate_list_UUEdL
#> 2 Sepal.Width  sepal.size aggregate_list_UUEdL
#> 3 Petal.Length petal.size aggregate_list_UUEdL
#> 4 Petal.Width  petal.size aggregate_list_UUEdL
juice(rec)
#> # A tibble: 150 × 3
#>    Species sepal.size petal.size
#>    <fct>        <dbl>      <dbl>
#>  1 setosa        17.8       0.28
#>  2 setosa        14.7       0.28
#>  3 setosa        15.0       0.26
#>  4 setosa        14.3       0.3 
#>  5 setosa        18         0.28
#>  6 setosa        21.1       0.68
#>  7 setosa        15.6       0.42
#>  8 setosa        17         0.3 
#>  9 setosa        12.8       0.28
#> 10 setosa        15.2       0.15
#> # ℹ 140 more rows